[Bronze II] Title: Hashing, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.cc
@@ -9,7 +9,7 @@ int main(void) {
 	long long hash = 0;
 	cin >> l;
 	for (; i < l; i++) {
-		static int coef = 1;
+		static long long coef = 1;
 		char letter;
 		cin >> letter;
 		long long num = letter - 'a' + 1;
